+++ rust/src/player/mod.rs
@@ -9,7 +9,7 @@ use std::{path::PathBuf, time::Duration, time::Instant};
 use url::Url;
 
 use crate::playback_view::{DragPhase, PlaybackViewDelegate};
-use crate::video::Video;
+use crate::video::VideoBackend;
 
 mod history;
 mod inspector;
@@ -62,7 +62,7 @@ pub(crate) fn bind_keys(cx: &mut App) {
 }
 
 pub(crate) struct Player {
-    video: Option<Video>,
+    video: Option<VideoBackend>,
     history: HistoryData,
     current_media_path: Option<PathBuf>,
     title: String,
@@ -180,7 +180,7 @@ impl Player {
             return;
         };
 
-        match Video::open(&url) {
+        match VideoBackend::open(&url) {
             Ok(video) => {
                 self.video = Some(video);
                 self.history.record(&path, title.clone());
